medicine authorization

+++ abl/medicine-abl.js
@@ -8,7 +8,7 @@ async function createMedicineAbl(req, res) {
 			return res.status(404).json({ message: "MedsTaker not found" });
 		}
 		if (medsTaker.supervisor !== req.userId) {
-			return res.status(403).json({ message: "Forbidden" });
+			return res.status(403).json({ message: "User is not authorized" });
 		}
 
 		const newMedicine = await MedicineDAO.createMedicine(req.body);
@@ -22,12 +22,18 @@ async function createMedicineAbl(req, res) {
 async function getMedicineAbl(req, res) {
 	try {
 		const medicine = await MedicineDAO.getMedicine(req.params.id);
-		//TODO authorization
-
 		if (!medicine) {
 			return res.status(404).json({ message: "Medicine not found" });
 		}
 
+		const medsTaker = await MedsTakerDAO.GetMedsTaker(medicine.medsTaker);
+		if (!medsTaker) {
+			return res.status(404).json({ message: "MedsTaker not found" });
+		}
+		if (medsTaker.supervisor !== req.userId) {
+			return res.status(403).json({ message: "User is not authorized" });
+		}
+		//TODO check if unit exists
 		res.status(200).json(medicine);
 	} catch (error) {
 		res.status(500).json({ message: error.message });
@@ -36,8 +42,16 @@ async function getMedicineAbl(req, res) {
 
 async function getMedicineByMedsTakerAbl(req, res) {
 	try {
+		const medsTaker = await MedsTakerDAO.GetMedsTaker(req.params.medsTakerId);
+		if (!medsTaker) {
+			return res.status(404).json({ message: "MedsTaker not found" });
+		}
+		if (medsTaker.supervisor !== req.userId) {
+			return res.status(403).json({ message: "User is not authorized" });
+		}
+		//TODO check if unit exists
 		const medicines = await MedicineDAO.getMedicineByMedsTaker(req.params.medsTakerId);
-		//TODO authorization
+
 		res.status(200).json(medicines);
 	} catch (error) {
 		res.status(500).json({ message: error.message });
@@ -50,7 +64,15 @@ async function deleteMedicineAbl(req, res) {
 		if (!medicine) {
 			return res.status(404).json({ message: "Medicine not found" });
 		}
-		//TODO: authorization
+
+		const medsTaker = await MedsTakerDAO.GetMedsTaker(medicine.medsTaker);
+		if (!medsTaker) {
+			return res.status(404).json({ message: "MedsTaker not found" });
+		}
+		if (medsTaker.supervisor !== req.userId) {
+			return res.status(403).json({ message: "User is not authorized" });
+		}
+		//TODO check if unit exists
 		const deletedMedicine = await MedicineDAO.deleteMedicine(req.params.id);
 		res.status(200).json(deletedMedicine);
 	} catch (error) {
@@ -65,8 +87,15 @@ async function updateMedicineAbl(req, res) {
 			return res.status(404).json({ message: "Medicine not found" });
 		}
 
+		const medsTaker = await MedsTakerDAO.GetMedsTaker(medicine.medsTaker);
+		if (!medsTaker) {
+			return res.status(404).json({ message: "MedsTaker not found" });
+		}
+		if (medsTaker.supervisor !== req.userId) {
+			return res.status(403).json({ message: "User is not authorized" });
+		}
+		//TODO check if unit exists
 		const updatedMedicine = await MedicineDAO.updateMedicine(req.params.id, req.body);
-		//TODO authorization
 		res.status(200).json(updatedMedicine);
 	} catch (error) {
 		res.status(500).json({ message: error.message });
